requirements.txt
mysql-connector-python
mysqlclient
SQLAlchemy
python-dateutil
requests
requests_oauthlib
matplotlib
numpy